* Fix to #1097623: 64-bit instrument selection
git-svn-id: http://svn.code.sf.net/p/rosegarden/code/trunk@6239 cf6a83e7-f017-4305-b0b3-678bde24f230

+++ sound/MappedDevice.cpp
@@ -205,7 +205,7 @@ operator>>(QDataStream &dS, MappedDevice &mD)
 QDataStream&
 operator<<(QDataStream &dS, MappedDevice *mD)
 {
-    dS << mD->size();
+    dS << (int)mD->size();
 
     for (MappedDeviceIterator it = mD->begin(); it != mD->end(); it++)
         dS << (*it);
@@ -228,7 +228,7 @@ operator<<(QDataStream &dS, MappedDevice *mD)
 QDataStream&
 operator<<(QDataStream &dS, const MappedDevice &mD)
 {
-    dS << mD.size();
+    dS << (int)mD.size();
 
     for (MappedDeviceConstIterator it = mD.begin(); it != mD.end(); it++)
         dS << (*it);
